ecs: Add multi iterator to `composition.zig`
The `MultiComponentIterator` allows iterating over multiple components,
and is made of a composition of `ComponentIterator`s. This is useful
when you want to iterate over multiple components at the same time.
In the future, we would probably abstract this out into the `World` so
the composition only has to wordd about the components it contains.